Support dynamic DNS in Froxlor
Customers can update the IP addresses of their domains via a simple HTTP
request.  This must be enabled for the domain before it is available. IPv4 and
IPv6 is supported. Multiple domains in one call are supported.

The `updateip.php` script uses HTTP Basic Authentication. With mod_php and
fcgid no configuration is required but for php-fpm the option '-pass-header
Authorization / CGIPassAuth On' must be enabled for the PHP configuration of
Froxlor's vhost.