Add the posibility to create new mailboxes with first- and lastname
This patch provides the positbility to create a new mailbox with a firstname and a
lastname. This is very useful if you want to connect these mailboxes with a groupware.
For the case that the feature is not wanted: There was added a setting for this feature.
Default status of this feature is disabled.