This commit fixes that dovecot and postfix were getting incorrect information about an email-user with a IDN domain.
With dovecot's password_query selecting username as email, we were overwriting the puny-encoded domain a client delivered
with the real unicode one (see https://wiki2.dovecot.org/PasswordDatabase/ExtraFields ), which currently doesn't create problems
(because e.g. we are selecting the `home` directory independently of the username and take it from the DB), but it is incorrect
nontheless.

In postfix's virtual_sender_permission however, this was creating problems when the reject_sender_login_mismatch sender restriction
was being used, as the login-name of the user (punycode) was not equal to the login-name (unicode) that is allowed to send with
that sender.